Revert "Fix the format string warnings."
This reverts commit 7350326dc3813807432a1f94877ddbec5883bdc6.

+++ Raamattu/MAChapterListViewController.m
@@ -106,7 +106,7 @@
     
     [_chapters removeAllObjects];
     for (NSUInteger i=1, max=_book.chapterIndex.chapterCount; i <= max; i++) {
-        NSString *chapterNumber = [NSString stringWithFormat:@"%lu", (unsigned long)i];
+        NSString *chapterNumber = [NSString stringWithFormat:@"%i", i];
         [_chapters addObject:chapterNumber];
     }
